requirements.txt
requests==2.25.1
singer-python==5.12.1
singer-tools==0.4.1
wrapt==1.12.1
pendulum==2.1.2